research/requirements.txt
openai>=1.60.0,<2.0.0
pydantic>=2.9,<3.0
python-dotenv>=1.0.0,<2.0.0
